+++ routes/api/cars.js
@@ -1,4 +1,3 @@
-require("dotenv").config();
 const express = require('express');
 const router = express.Router();
 const config = require('../../config.json')
@@ -6,17 +5,10 @@ const carsDB = require('../../models/schemas/cars')
 const { apiAuthCheck } = require('../../middleware/authentication/apiAuthentication');
 const { checkOrigin } = require('../../middleware/authentication/checkOrigin');
 
-const AWS = require('aws-sdk')
+const { s3 } = require('../../db');
 const multer = require('multer');
 
-// Configure AWS SDK
-AWS.config.update({
-    accessKeyId: process.env.AWS_ACCESS_KEY,
-    secretAccessKey: process.env.AWS_SECRET_KEY,
-    region: process.env.AWS_REGION
-});
 
-const s3 = new AWS.S3();
 
 // Multer upload middleware
 const upload = multer({
@@ -38,7 +30,7 @@ router.post('/add', upload.array('images'), apiAuthCheck, checkOrigin, async (re
         const uploadedImageUrls = await Promise.all(images.map(async (image) => {
             const uploadParams = {
                 Bucket: 'reubz',
-                Key: `${carData.brand.toLowerCase()}-${carData.name.toLowerCase()}/${Math.floor(Math.random() * 100)}-${image.originalname.toLowerCase()}`, // You can set a different key as per your requirement
+                Key: `cars/${carData.brand.toLowerCase()}-${carData.name.toLowerCase()}/${Math.floor(Math.random() * 100)}-${image.originalname.toLowerCase()}`, // You can set a different key as per your requirement
                 Body: image.buffer, // Image data
                 // ACL: 'public-read', // Make the uploaded image public
                 ContentType: image.mimetype // Mime type of the image